requirements.txt
Flask==1.0.2
Flask-MySQL==1.4.0
PyJWT==1.6.4
gunicorn==19.9.0
flask-MySQL
flaskext.mysql